requirements.txt
Jinja2==3.1.2
openpyxl==3.1.2
pandas==1.5.3
python-dotenv==1.0.0